requirements.txt
streamlit>=1.28.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
openai>=1.50.0
boto3>=1.34.0
pinecone-client>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0